fix: Move DeletePredicate types to data_types2
