+++ data_types/src/lib.rs
@@ -15,7 +15,6 @@ pub mod chunk_metadata;
 pub mod consistent_hasher;
 mod database_name;
 pub mod database_rules;
-pub mod delete_predicate;
 pub mod error;
 pub mod job;
 pub mod names;
