+++ data_types2/Cargo.toml
@@ -2,11 +2,12 @@
 name = "data_types2"
 version = "0.1.0"
 edition = "2021"
-description = "Shared data types in the Iox NG architecture"
+description = "Shared data types in the IOx NG architecture"
 
 [dependencies]
 data_types = { path = "../data_types" }
 influxdb_line_protocol = { path = "../influxdb_line_protocol" }
+ordered-float = "3"
 schema = { path = "../schema" }
 sqlx = { version = "0.5", features = ["runtime-tokio-rustls", "postgres", "uuid"] }
 uuid = { version = "0.8", features = ["v4"] }
